幸运大抽奖--抽奖活动奖励表修改

+++ jmp/mp-service/src/main/java/com/vpu/mp/service/pojo/shop/market/lottery/prize/LotteryPrizeParam.java
@@ -18,9 +18,9 @@ public class LotteryPrizeParam {
     private Integer lotteryNumber;
     private Integer awardTimes;
     private Integer integralScore;
-    private Byte    couponId;
-    private Byte    prdId;
-    private Byte    prdKeepDays;
+    private Integer    couponId;
+    private Integer    prdId;
+    private Short    prdKeepDays;
     private String  lotteryDetail;
     private String  iconImgsImage;
     private String  iconImgs;
